# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EAAPA – East Africa Agripreneurs Alliance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List investment opportunities
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Opportunities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
    post:
      operationId: submitOpportunity
      tags: [opportunities]
      summary: Submit an investment idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunity"
      responses:
        "201":
          description: Opportunity submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /projects:
    get:
      operationId: listProjects
      tags: [network]
      summary: List collaboration projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [network]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/join:
    post:
      operationId: joinProject
      tags: [network]
      summary: Join a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /forum/threads:
    get:
      operationId: listForumThreads
      tags: [community]
      summary: List forum threads
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Forum threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumThread"
    post:
      operationId: createForumThread
      tags: [community]
      summary: Create a forum thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumThread"
      responses:
        "201":
          description: Thread created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumThread"
...
components:
  schemas:
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [investment, funding, value_chain, grant, accelerator, incubator]
        sector:
          type: string
        roi:
          type: string
        fundingAmount:
          type: number
        currency:
          type: string
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [open, closing_soon, closed]
        tags:
          type: array
          items:
            type: string
      required: [id, title, description, type, sector, status]
    CreateOpportunity:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        sector:
          type: string
        roi:
          type: string
        fundingAmount:
          type: number
      required: [title, description, type, sector]
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        leader:
          type: string
        leaderId:
          type: integer
        status:
          type: string
          enum: [planning, active, completed]
        commodity:
          type: string
        region:
          type: string
        startDate:
          type: string
          format: date
        memberCount:
          type: integer
        isJoined:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required: [id, title, description, leader, leaderId, status, memberCount, isJoined]
    CreateProject:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        commodity:
          type: string
        region:
          type: string
        startDate:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
      required: [title, description]
    ForumThread:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        authorId:
          type: integer
        replies:
          type: integer
        views:
          type: integer
        createdAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
      required: [id, title, content, category, author, authorId, replies, views, createdAt]
    CreateForumThread:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, content, category]
...